tweeking DDA

+++ software/WastelandWheeler/Assets/Scripts/DDA/IDiffcultyAdjuster.cs
@@ -5,5 +5,5 @@ public interface IDiffcultyAdjuster
 {
     // method to affect the game making it easier or hard
     // negative amounts make the game easier, positive make it harder
-    void ChangeDifficulty(float amount);
+    void ChangeDifficulty(int amount);
 }
